requirements.txt
gitpython>=3.1.40
PyYAML>=6.0
psutil>=5.9.0
requests>=2.28.0
